Creación puntos 1 y 2

+++ punto-1/controller/calculationController.php
@@ -10,8 +10,8 @@ $message = "";
 if(isset($_POST["first_number"]) &&
    isset($_POST["second_number"]) &&
    isset($_POST["operation_type"])) {
-    $firstNumber = intval($_POST["first_number"]);
-    $secondNumber = intval($_POST["second_number"]);
+    $firstNumber = floatval($_POST["first_number"]);
+    $secondNumber = floatval($_POST["second_number"]);
     $operationType = $_POST["operation_type"];  
     
     $result = calculate($firstNumber, $secondNumber, $operationType);
